Move standard article grid into its own component
